// src/app/servicesahmed/models/offre.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Offre {
  createdAt?: string;    // <-- added this line for createdAt
  createdBy?: number;
  description?: string;
  id?: number;
  imageUrl?: string;
  skills?: string;
  title?: string;
}